LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/credentialless/cache-storage.tentative.https-expected.txt

CONSOLE MESSAGE: TypeError: constructor is not a function. (In 'constructor(coep_none)', 'constructor' is undefined)

Harness Error (FAIL), message = TypeError: constructor is not a function. (In 'constructor(coep_none)', 'constructor' is undefined)
